% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.R
\name{calculate_PM}
\alias{calculate_PM}
\title{Calculate performance metric}
\usage{
calculate_PM(.mse, PMs, all_sims = FALSE)
}
\arguments{
\item{.mse}{MSE object}

\item{PMs}{Character vector of the performance metric functions}

\item{all_sims}{Logical, whether to report the value for individual simulations (TRUE), or averaged over all simulations (FALSE)}
}
\value{
If \code{all_sims = TRUE} a matrix, otherwise a data frame
}
\description{
Wrapper function that conveniently calculates the values for a set of performance metrics
}
