% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{make_hake_MP}
\alias{make_hake_MP}
\title{Make model-based management procedure}
\usage{
make_hake_MP(HCR_fn)
}
\arguments{
\item{HCR_fn}{A harvest control rule function that converts assessment output into catch advice, excluding hyper-rule}

\item{delta}{Vector length 2, minimum and maximum change in CBA (applied after lambda)}
}
\value{
A MP function that generates the CBA
}
\description{
Creates a management procedure from an assessment and harvest control rule
}
