% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om.R
\name{make_hake_OM}
\alias{make_hake_OM}
\alias{make_hake_OM_AMAC}
\title{Create hake operating model from ADMB estimation model}
\usage{
make_hake_OM(
  dat = "admb-assessment/datos_14021.dat",
  rep = "admb-assessment/modelo_140.rep",
  Name = "Merluza comun - OM 1",
  nsim = 100,
  proyears = 36,
  AC = 0.65,
  Perr = 0.5,
  tv_maturity = FALSE
)

make_hake_OM_AMAC(
  jjdir = "admb-assessment/om2",
  model = "mc_0.0",
  Name = "Merluza comun - OM 2",
  nsim = 100,
  proyears = 36,
  AC = 0.65,
  Perr = 0.5
)
}
\arguments{
\item{dat}{Filename of ADMB data file}

\item{rep}{Filename of ADMB report file}

\item{Name}{Character, name of operating model to insert to \code{OM@Name}}

\item{nsim}{Integer, number of simulations in the operating model}

\item{proyears}{Integer, number of projection years in the operating model}

\item{AC}{Numeric, lag-1 autocorrelation for the simulated recruitment deviations in the projection}

\item{Perr}{Numeric, lognormal standard deviation for the simulated recruitment deviations in the projection}

\item{tv_maturity}{Logical, whether to use time-varying maturity in the operating model (specified in the ADMB data file)}

\item{jjdir}{Directory of the JJM assessment}

\item{model}{Character string for name of model}
}
\value{
\link[MSEtool:OM-class]{MSEtool::OM} object
}
\description{
Reads in ADMB data and report files to generate an openMSE operating model
}
