% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj-figures.R
\name{plot_array}
\alias{plot_array}
\title{Plot time series figures}
\usage{
plot_array(
  MSE,
  names = "",
  sims,
  type = c("SB", "B_BMSY", "F_FMSY", "CBA", "Index", "Catch", "R")
)
}
\arguments{
\item{MSE}{Either a list of \link[MSEtool:MSE-class]{MSEtool::MSE} objects or single MSE object}

\item{names}{Character vector for the name of the operating model}

\item{sims}{Integer vector for a subset of simulations to plot}

\item{type}{Character for the state variable to plot}
}
\value{
A ggplot2 object
}
\description{
Plots state variables for each operating model and management procedure
}
