% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.R
\name{plot_table}
\alias{plot_table}
\title{Performance metric table}
\usage{
plot_table(
  probs_dat,
  names,
  ncol = NULL,
  relative_max = FALSE,
  scale_0_1 = FALSE,
  sort_by = "decreasing",
  mp_order = NULL,
  satisficed = NULL,
  alpha = 0.6,
  do_approx = FALSE
)
}
\arguments{
\item{probs_dat}{List of data frames containing performance measures}

\item{names}{Character vector of operating model names corresponding to \code{probs_dat}}

\item{ncol}{Integer, optional number of columns.}

\item{relative_max}{Make the plot have each column use a relative maximum. If
\code{scale_0_1} is used, this will be ignored}

\item{scale_0_1}{Scale each column from 0 to 1, so that the colours in each
column are fully represented}

\item{sort_by}{show values in decreasing or increasing format}

\item{mp_order}{Optional hardcoded MP order}

\item{satisficed}{An optional named numeric vector. The names correspond to
the performance metrics and the values correspond to the values above which
(\code{>}) the cells will be outlined as "satisficed".}

\item{alpha}{Transparency of underlying colour.}

\item{do_approx}{Logical. If \code{TRUE}, values greater than 0.99 are replaced with ">0.99" and less than 0.01 are replaced
"<0.01".}
}
\value{
A ggplot2 object
}
\description{
Create coloured performance metric table in ggplot2 for a set of operating models. Code
taken from \url{https://github.com/pbs-assess/ggmse/blob/master/R/figures-tigure.R}
}
