% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.R
\name{plot_tradeoff}
\alias{plot_tradeoff}
\title{Plot tradeoff figure}
\usage{
plot_tradeoff(
  pm_df_list,
  xvar,
  yvar,
  custom_pal = NULL,
  mp = NULL,
  mp_ref = NULL,
  nudge_x = 0,
  nudge_y = 0.05
)
}
\arguments{
\item{pm_df_list}{A named list of performance metric data frames. The names will be used as the plot labels.}

\item{xvar}{The performance metric for the x axis (as character).}

\item{yvar}{The performance metric for the y axis (as character).}

\item{custom_pal}{An optional custom color palette. Should be a named character vector.}

\item{mp}{An optional character vector of MPs to include. By default includes all.}

\item{mp_ref}{Character vector of reference management procedures}

\item{nudge_x}{How far to nudge the labels left/right from the x-value for \link[ggplot2]{geom_text}}

\item{nudge_y}{How far to nudge the labels up/down from the y-value for \link[ggplot2]{geom_text}}
}
\value{
A ggplot2 object
}
\description{
Create figure to compare management procedures across 2 performance metrics. Code is copied from
\url{https://github.com/pbs-assess/ggmse/blob/master/R/figures-bivariate.R#L26}
}
