% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj-figures.R
\name{plot_yield_curve}
\alias{plot_yield_curve}
\title{Yield curve}
\usage{
plot_yield_curve(
  MSE_list,
  names,
  Fvec = c(1e-08, seq(0.01, 3, 0.01)),
  Fmaxplot = 0.8,
  yr.ind = MSE_list[[1]]@nyears,
  by_OM = FALSE
)
}
\arguments{
\item{MSE_list}{List of \link[MSEtool:MSE-class]{MSEtool::MSE} objects}

\item{names}{Character vector for the name of each operating model}

\item{Fvec}{Numeric vector for values of fishing mortality to calculate the yield curve}

\item{Fmaxplot}{Numeric, axis limit for F in the ggplot figure}

\item{yr.ind}{Integer, year of which to obtain the biological parameters and selectivity for the yield curve}

\item{by_OM}{Logical, whether to plot individual panels for each operating model}
}
\value{
ggplot object
}
\description{
Compare yield curve from set of operating models
}
