% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCM2Assess.R
\name{RCM2Assess}
\alias{RCM2Assess}
\title{Make hake assessment model}
\usage{
RCM2Assess(RCModel)
}
\arguments{
\item{RCModel}{\link[SAMtool:RCModel-class]{SAMtool::RCModel} object returned by \code{\link[SAMtool:RCM]{SAMtool::RCM()}}. The estimation model will be converted to an assessment function}
}
\value{
A function that can be used in a management procedure
}
\description{
Convert an operating model conditioned by RCM (TMB estimation model) into an assessment model for simulation testing of management procedures
}
