% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\docType{data}
\name{SPHenv}
\alias{SPHenv}
\title{Environment for storing hake assessment object}
\format{
An object of class \code{environment} of length 2.
}
\usage{
SPHenv
}
\description{
An environment that stores variables for the hake assessment.
}
\keyword{datasets}
